perf: optimise changelogs
