+++ .gitignore
@@ -22,6 +22,7 @@ terraform/**/.env
 terraform/**/**/*.tfstate*
 /.cache
 /test-results
+.pnpm/
 
 # Sentry Config File
 .env.sentry-build-plugin
