+++ src/routes/changelog/[[page]]/+page.ts
@@ -1,11 +1,10 @@
 import { redirect } from '@sveltejs/kit';
-import { CHANGELOG_DEPENDENCY, getAllChangelogEntries } from '../utils';
+import { CHANGELOG_DEPENDENCY, changelogCount, changelogs } from '../utils';
 
 const PER_PAGE = 5;
 
 export const entries = async () => {
-    const entries = await getAllChangelogEntries();
-    const totalPages = Math.ceil(entries.length / PER_PAGE);
+    const totalPages = Math.ceil(changelogCount / PER_PAGE);
     return Array.from({ length: totalPages }, (_, i) => {
         return {
             page: (i + 1).toString()
@@ -22,11 +21,9 @@ export const load = async ({ depends, params }) => {
 
     const page = parseInt(params.page || '1', 10);
 
-    const entries = await getAllChangelogEntries();
-
     return {
         page,
-        entries: entries.slice(0, page * PER_PAGE),
-        nextPage: page < Math.ceil(entries.length / PER_PAGE) ? page + 1 : null
+        entries: changelogs.slice(0, page * PER_PAGE),
+        nextPage: page < Math.ceil(changelogCount / PER_PAGE) ? page + 1 : null
     };
 };
