+++ src/routes/changelog/entry/[entry]/+page.ts
@@ -1,21 +1,14 @@
 import { error } from '@sveltejs/kit';
-import { getAllChangelogEntries } from '../../utils.js';
+import { changelogs } from '../../utils.js';
 import type { EntryGenerator } from './$types.js';
 
-export const entries: EntryGenerator = async () => {
-    const entries = await getAllChangelogEntries();
-    return entries.map((entry) => {
-        return {
-            entry: entry.slug
-        };
-    });
-};
+export const entries: EntryGenerator = async () =>
+    changelogs.map((entry) => ({
+        entry: entry.slug
+    }));
 
 export const load = async ({ params }) => {
-    const entries = await getAllChangelogEntries();
-    const entry = entries.find((entry) => {
-        return entry.slug === params.entry;
-    });
+    const entry = changelogs.find((entry) => entry.slug === params.entry);
 
     if (!entry) {
         error(404, 'Not found');
